refactor(slic-tools): rename aws util to dynamo
