+++ packages/checklist-service/services/checklists/entries/entries.js
@@ -3,7 +3,7 @@ const {
   UpdateCommand
 } = require('@aws-sdk/lib-dynamodb')
 const { v4: uuid } = require('uuid')
-const { dynamoDocClient } = require('slic-tools/aws')
+const { dynamoDocClient } = require('slic-tools/dynamo')
 
 const { createMetricsLogger, Unit } = require('aws-embedded-metrics')
